Move main method to calculate_next_version.py
